pyproject.toml
[project]
name = "squish"
version = "0.1.0"
description = ""
authors = [
    {name = "Septicuss",email = "40741898+Septicuss@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "invoke (>=2.2.0,<3.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
coverage = "^7.8.0"

